Make filename relative to home dir.
I would like to also use ":.", but this should not remove "~/" when
in ~, because I think that `~/` is a prefix worth keeping and it should
get "./" prepended probably to indicate "current directory".

Apart from that, the filenames' leading path segment (before any slash)
are currently highlighted in red - I could not find where that happens.
It should consider "./" and "~/" to be valid prefixes, too.